requirements.txt
google-cloud-aiplatform
google-genai
google-cloud-storage
google-cloud-bigquery
pydantic
python-dotenv
httpx
pytest
numpy
streamlit
jsonschema
pyyaml
requests
